requirement.txt
Flask==2.2.2
mysql-connector-python==8.0.33
requests==2.28.2
beautifulsoup4==4.11.1
